SSL Cleaned up. I belive this should work
