+++ src/main/java/no/ntnu/idata2304/group1/server/network/TCPListener.java
@@ -1,13 +1,7 @@
 package no.ntnu.idata2304.group1.server.network;
 
 import java.io.Closeable;
-import java.io.File;
-import java.io.FileInputStream;
 import java.io.IOException;
-import java.io.InputStream;
-import java.security.KeyStore;
-import java.security.SecureRandom;
-import javax.net.ssl.KeyManagerFactory;
 import javax.net.ssl.SSLContext;
 import javax.net.ssl.SSLServerSocket;
 import javax.net.ssl.SSLServerSocketFactory;
@@ -24,26 +18,15 @@ public class TCPListener extends Thread implements Closeable {
     private SSLServerSocket socket;
 
     /**
-     * Creates a new TCP listener
+     * Creates a new TCP listener.
      * 
      * @param port The port to listen on
+     * @param keyStorePath The path to the keystore
+     * @param keyStorePassword The password to the keystore
      * @throws IOException if the socket fails to connect or the keystore is not found
      */
     public TCPListener(int port, String keyStoreName, String keyStorePassword) throws IOException {
-        SSLContext context = createSSLContext(keyStoreName, keyStorePassword);
-        if (context == null) {
-            throw new IOException("Could not create SSL context");
-        }
-        SSLServerSocketFactory factory = context.getServerSocketFactory();
-        this.socket = (SSLServerSocket) factory.createServerSocket(port);
-        this.socket.setEnabledCipherSuites(factory.getDefaultCipherSuites());
-        String[] supported = this.socket.getSupportedProtocols();
-        this.socket.setEnabledProtocols(supported);
-    }
-
-    public TCPListener(int port) throws IOException {
-        SSLContext context = createSSLContext(
-                this.getClass().getResource("serverKeys").getPath().replace("%20", ""), "123");
+        SSLContext context = SeverSSLKeyFactory.createSSLContext(keyStoreName, keyStorePassword);
         if (context == null) {
             throw new IOException("Could not create SSL context");
         }
@@ -72,58 +55,5 @@ public class TCPListener extends Thread implements Closeable {
     @Override
     public void close() throws IOException {
         socket.close();
-    }
-
-    /**
-     * Creates a SSL context from the keystore This code is based on the example from this page:
-     * https://gpotter2.github.io/tutos/en/sslsockets
-     * 
-     * @param keyStorePath
-     * @param keyStorePassword
-     * @return
-     */
-    private SSLContext createSSLContext(String keyStorePath, String keyStorePassword) {
-        SSLContext ctx = null;
-        try {
-            KeyStore keyStore = KeyStore.getInstance("pkcs12");
-            try (InputStream kstore = new FileInputStream(keyStorePath)) {
-                keyStore.load(kstore, keyStorePassword.toCharArray());
-            }
-            KeyManagerFactory kmf =
-                    KeyManagerFactory.getInstance(KeyManagerFactory.getDefaultAlgorithm());
-            kmf.init(keyStore, "".toCharArray());
-            ctx = SSLContext.getInstance("TLS");
-            ctx.init(kmf.getKeyManagers(), null, SecureRandom.getInstanceStrong());
-        } catch (Exception e) {
-            System.out.println("Error creating SSL context");
-        }
-        return ctx;
-    }
-
-    /**
-     * Check that the file actually exists
-     * 
-     * @param path The path to the file
-     * @return true if the file exists, false otherwise
-     */
-    // TODO: Move this to a utility class
-    // TODO: Implement a check for trying to fetch the server keystore
-    public static boolean checkServerKeyExist(String path) {
-        boolean returnValue = false;
-        if (path != null && !path.isBlank()) {
-            File file = new File(path);
-            returnValue = file.exists();
-        }
-        return returnValue;
-    }
-
-    /**
-     * Check that the default server keystore exists
-     * 
-     * @return true if the file exists, false otherwise
-     */
-    public static boolean checkServerKeyExist() {
-        return checkServerKeyExist(
-                TCPListener.class.getResource("serverKeys").getPath().replace("%20", " "));
     }
 }
